Fixed leak, added safety comments, code style
The Leak: I wasn't dropping the wdesc and desc in the case that cmpxchng
failed. I thought that the cmpxchng would always succeed, as the main
function I was using to test is single threaded. However, when I tracked the function calls
happening, I realized that it was failing because I was using the weak
version and I need to dealloc the allocs from the failed cmpxchngs.