(Knock on wood) Implemented reclamation, no ub
The main issue with reclamation for the vector is descs/wdescs. The
buffers are always deallocated when the vector gets dropped. Descs
themselves are not that hard, just retire it when it get's compare and
swapped out. The problem is wdescs. You can't retire a desc's wdesc
when you retire the desc because there could be a thread reading the
desc who then loads the *retired* wdesc. This is against the safety
contract for the retire method in the haphazard docs. I was
inadvertently doing this earlier, and when I reread the docs, I
realized this was the source of my UB.

The way I've no implemented reclamation is by implementing Drop for
descs. When a desc is dropped (once it is deemed hazard-free), it's
wdesc is dropped. Because Drop takes &mut self, in theory, there can
be no references into the desc for reading the wdesc. I'm not sure if
the haphazard library actually checks it, or the compiler defers
deallocation until there are no references or if neither happens.
Because of this, I have made sure that wdesc references never outlive
the reference to the desc which contains them. I tried to ensure this
with lifetimes but I wasn't able to write an adequate solution.

There is a special case for deallocation of wdescs. When a wdesc gets
swapped out by `complete_write`, there can be no new loads to it
because it is impossible to access without a reference from before it
was swapped; there is no descriptor pointing to it anymore. Thus, we
can just retire the swapped out wdesc.

Note: need to add safety comments!!!